moved to .NET 5 added tests for DayOffRepository

+++ src/EMS.Gateway.API/DAL/Repositories/PositionsRepository.cs
@@ -7,12 +7,12 @@ using System.Threading.Tasks;
 
 namespace EMS.Gateway.API.DAL.Repositories
 {
-    public class PositionsRepository: BaseRepository, IPositionsRepository
-	{
-		public PositionsRepository(IApplicationDbContext context): base(context) { }
+    public class PositionsRepository : BaseRepository, IPositionsRepository
+    {
+        public PositionsRepository(IApplicationDbContext context) : base(context) { }
 
-		public async Task<int> AddAsync(Position position)
-		{
+        public async Task<int> AddAsync(Position position)
+        {
             if (position is null)
             {
                 throw new ArgumentNullException("Position cannot be empty");
@@ -27,11 +27,11 @@ namespace EMS.Gateway.API.DAL.Repositories
                 throw new DbUpdateException("Cannot add position with non exists team");
             }
             _context.Positions.Add(position);
-			return await _context.SaveChangesAsync();
-		}
+            return await _context.SaveChangesAsync();
+        }
 
-		public async Task<int> UpdateAsync(Position position)
-		{
+        public async Task<int> UpdateAsync(Position position)
+        {
             if (position is null)
             {
                 throw new ArgumentNullException("Position cannot be empty");
@@ -40,33 +40,37 @@ namespace EMS.Gateway.API.DAL.Repositories
             {
                 throw new ArgumentNullException("Position name cannot be empty");
             }
-            if(position.TeamId == 0 || _context.Teams.FirstOrDefault(t=> t.Id == position.TeamId) is null ||
+            if (position.TeamId == 0 || _context.Teams.FirstOrDefault(t => t.Id == position.TeamId) is null ||
                 _context.Teams.First(t => t.Id == position.TeamId) != position.Team)
             {
                 throw new DbUpdateException("Cannot add position with non exists team");
             }
             _context.Positions.Update(position);
-			return await _context.SaveChangesAsync();
-		}
+            return await _context.SaveChangesAsync();
+        }
 
-		public async Task<int> DeleteAsync(Position position)
-		{
-            if(_context.Teams.Any(e => e.Id == position.TeamId))
+        public async Task<int> DeleteAsync(Position position)
+        {
+            if (_context.Teams.Any(e => e.Id == position.TeamId))
             {
                 throw new DbUpdateException("Cannot delete position related to team");
             }
+            if (_context.Staff.Any(e => e.PositionId == position.Id))
+            {
+                throw new DbUpdateException("Cannot delete position related to staff");
+            }
             _context.Positions.Remove(position);
-			return await _context.SaveChangesAsync();
-		}
+            return await _context.SaveChangesAsync();
+        }
 
-		public Position Get(long id)
-		{
-			return _context.Positions.FirstOrDefault(p => p.Id == id);
-		}
+        public Position Get(long id)
+        {
+            return _context.Positions.FirstOrDefault(p => p.Id == id);
+        }
 
-		public IQueryable<Position> GetAll()
-		{
+        public IQueryable<Position> GetAll()
+        {
             return _context.Positions.Select(e => e);
-		}
-	}
+        }
+    }
 }
